// CottageBooking/frontend/src/app/models/reservations.ts
export class Reservation {
    vikendica: string = "";       
    turista: string = "";         
    datumPocetka: string = "";
    datumKraja: string = "";
    brojOsoba: number = 1;
    status: string = "na čekanju";
    ocena?: number;               
    komentar?: string;   
    datum:string="";
    zahtevi?:string;  
    komentarOdbijanja?:string;     
}